#5820 fix code review findings
